<feat>(comm): added 'enums.py' file
