+++ shippa_order/merchants/serializers.py
@@ -1,36 +1,10 @@
-from types import MappingProxyType
 from datetime import datetime
 
 from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
 
 from merchants.models import Merchant, Menu
-
-
-def _available_days(business_day):
-    days = [(64, 'SUN', ), (32, 'MON', ), (16, 'TUE', ), (8, 'WED', ),
-            (4, 'THU', ), (2, 'FRI', ), (1, 'SAT', ), ]
-    return [day for binary, day in days if binary & business_day]
-
-
-BINARY_OF_DAYS = MappingProxyType({
-    'SUN': int('1000000', 2),
-    'MON': int('0100000', 2),
-    'TUE': int('0010000', 2),
-    'WED': int('0001000', 2),
-    'THU': int('0000100', 2),
-    'FRI': int('0000010', 2),
-    'SAT': int('0000001', 2)
-})
-
-MAPPING_OF_AVAILABLE_DAYS = MappingProxyType({
-    business_day: _available_days(business_day)
-    for business_day in range(1, 128)
-})
-
-AVAILABLE_COUNTRY_ISO_LIST = ['KR', ]
-
-AVAILABLE_CURRENCIES = ['KRW', ]
+from merchants.enums import BinaryOfDays, MAPPING_OF_AVAILABLE_DAYS, AVAILABLE_COUNTRY_ISO_LIST, AVAILABLE_CURRENCIES
 
 
 def _validate_time_format(value):
@@ -45,7 +19,7 @@ def _convert_to_binary_for_business_days(value):
     if not isinstance(value, list):
         raise ValidationError
     try:
-        ret = sum([BINARY_OF_DAYS[day] for day in value])
+        ret = sum([BinaryOfDays[day].value for day in value])
     except KeyError:
         raise ValidationError
     return ret
